+++ internal/exporter/aws_guid.go
@@ -1,4 +1,4 @@
-package main
+package exporter
 
 import "strings"
