+++ internal/exporter/aws_guid_test.go
@@ -1,4 +1,4 @@
-package main
+package exporter
 
 import "testing"
