+++ internal/exporter/gcp_feed_test.go
@@ -1,4 +1,4 @@
-package main
+package exporter
 
 import (
 	"io/ioutil"
