+++ internal/exporter/testdata/aws_multi_item.rss
