+++ internal/exporter/testdata/gcp_resolved_then_update.atom
