+++ internal/exporter/testdata/gcp_update_resolved.atom
