+++ internal/exporter/worker.go
@@ -1,4 +1,4 @@
-package main
+package exporter
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 
 // startWorkers launches a monitor goroutine for each service and returns a
 // WaitGroup that waits for all workers to exit.
-func startWorkers(ctx context.Context, services []ServiceFeed) *sync.WaitGroup {
+func StartWorkers(ctx context.Context, services []ServiceFeed) *sync.WaitGroup {
 	var wg sync.WaitGroup
 	for _, svc := range services {
 		wg.Add(1)
